Moved `MessageAttributeValue` to where it is used
